+++ embedding-stores/langchain4j-community-neo4j/src/main/java/dev/langchain4j/community/store/embedding/neo4j/Neo4jEmbeddingStore.java
@@ -1,6 +1,20 @@
 package dev.langchain4j.community.store.embedding.neo4j;
 
-import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.*;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.DEFAULT_AWAIT_INDEX_TIMEOUT;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.DEFAULT_DATABASE_NAME;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.DEFAULT_EMBEDDING_PROP;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.DEFAULT_FULLTEXT_IDX_NAME;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.DEFAULT_IDX_NAME;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.DEFAULT_ID_PROP;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.DEFAULT_LABEL;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.DEFAULT_TEXT_PROP;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.EMBEDDINGS_ROW_KEY;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.METADATA;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.PROPS;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.SCORE;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.getRowsBatched;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.sanitizeOrThrows;
+import static dev.langchain4j.community.store.embedding.neo4j.Neo4jEmbeddingUtils.toEmbeddingMatch;
 import static dev.langchain4j.internal.Utils.getOrDefault;
 import static dev.langchain4j.internal.Utils.isNullOrEmpty;
 import static dev.langchain4j.internal.Utils.randomUUID;
@@ -23,8 +37,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
 import org.neo4j.cypherdsl.support.schema_name.SchemaNames;
 import org.neo4j.driver.Driver;
 import org.neo4j.driver.Session;
@@ -72,6 +86,7 @@ public class Neo4jEmbeddingStore implements EmbeddingStore<TextSegment> {
                         `vector.similarity_function`: 'cosine'
                     }}
                     """;
+    public static final String COLUMNS_NOT_ALLOWED_ERR = "There are columns not allowed in the search query: ";
 
     /* Neo4j Java Driver settings */
     private final Driver driver;
@@ -96,7 +111,7 @@ public class Neo4jEmbeddingStore implements EmbeddingStore<TextSegment> {
     private final String fullTextIndexName;
     private final String fullTextQuery;
     private final String fullTextRetrievalQuery;
-    private final boolean fullTextAutocreate;
+    private final boolean autoCreateFullText;
 
     /**
      * Creates an instance of Neo4jEmbeddingStore
@@ -114,10 +129,10 @@ public class Neo4jEmbeddingStore implements EmbeddingStore<TextSegment> {
      * @param awaitIndexTimeout the optional awaiting timeout for all indexes to come online, in seconds (default: 60s)
      * @param retrievalQuery    the optional retrieval query
      *                          (default: "RETURN properties(node) AS metadata, node.`idProperty` AS `idProperty`, node.`textProperty` AS `textProperty`, node.`embeddingProperty` AS `embeddingProperty`, score")
-     * @param fullTextIndexName the optional full-text index name, to perform a hybrid search (default: `fulltext`) 
-     * @param fullTextQuery the optional full-text index query, required if we want to perform a hybrid search                        
-     * @param fullTextRetrievalQuery the optional full-text retrieval query (default: {@param retrievalQuery})                        
-     * @param fullTextAutocreate if true, it will auto create the full-text index if not exists (default: false)                        
+     * @param fullTextIndexName the optional full-text index name, to perform a hybrid search (default: `fulltext`)
+     * @param fullTextQuery the optional full-text index query, required if we want to perform a hybrid search
+     * @param fullTextRetrievalQuery the optional full-text retrieval query (default: {@param retrievalQuery})
+     * @param autoCreateFullText if true, it will auto create the full-text index if not exists (default: false)
      */
     public Neo4jEmbeddingStore(
             SessionConfig config,
@@ -135,7 +150,7 @@ public class Neo4jEmbeddingStore implements EmbeddingStore<TextSegment> {
             String fullTextIndexName,
             String fullTextQuery,
             String fullTextRetrievalQuery,
-            boolean fullTextAutocreate) {
+            boolean autoCreateFullText) {
 
         /* required configs */
         this.driver = ensureNotNull(driver, "driver");
@@ -169,15 +184,20 @@ public class Neo4jEmbeddingStore implements EmbeddingStore<TextSegment> {
                 "RETURN properties(node) AS metadata, node.%1$s AS %1$s, node.%2$s AS %2$s, node.%3$s AS %3$s, score",
                 this.sanitizedIdProperty, sanitizedText, sanitizedEmbeddingProperty);
         this.retrievalQuery = getOrDefault(retrievalQuery, defaultRetrievalQuery);
+        //        retrievalQuery = getOrDefault(retrievalQuery, defaultRetrievalQuery);
+        //        this.retrievalQuery = sanitizeOrThrows(retrievalQuery, "retrievalQuery");
 
         this.notMetaKeys = new HashSet<>(Arrays.asList(this.idProperty, this.embeddingProperty, this.textProperty));
-        
+
         /* optional full text index */
-        this.fullTextAutocreate = fullTextAutocreate;
+        this.autoCreateFullText = autoCreateFullText;
         this.fullTextIndexName = getOrDefault(fullTextIndexName, DEFAULT_FULLTEXT_IDX_NAME);
         this.fullTextQuery = SchemaNames.sanitize(fullTextQuery).orElse(null);
+
         this.fullTextRetrievalQuery = getOrDefault(fullTextRetrievalQuery, this.retrievalQuery);
-        
+        //        fullTextRetrievalQuery = getOrDefault(fullTextRetrievalQuery, this.retrievalQuery);
+        //        this.fullTextRetrievalQuery = sanitizeOrThrows(fullTextRetrievalQuery, "fullTextRetrievalQuery");
+
         /* auto-schema creation */
         createSchema();
     }
@@ -303,11 +323,28 @@ public class Neo4jEmbeddingStore implements EmbeddingStore<TextSegment> {
                         "question", fullTextQuery));
             }
 
+            final String finalQuery = query;
+            final Set<String> columns = getColumnNames(session, query);
+            final Set<Object> allowedColumn = Set.of(textProperty, embeddingProperty, idProperty, SCORE, METADATA);
+
+            if (!allowedColumn.containsAll(columns) || columns.size() > allowedColumn.size()) {
+                throw new RuntimeException(COLUMNS_NOT_ALLOWED_ERR + columns);
+            }
+
             List<EmbeddingMatch<TextSegment>> matches =
-                    session.run(query, params).list(item -> toEmbeddingMatch(this, item));
+                    session.executeRead(tx -> tx.run(finalQuery, params).list(item -> toEmbeddingMatch(this, item)));
 
             return new EmbeddingSearchResult<>(matches);
         }
+    }
+
+    private static Set<String> getColumnNames(Session session, String query) {
+        // retrieve column names
+        final List<String> keys = session.run("EXPLAIN " + query).keys();
+        // when there are multiple variables with the same name, e.g. within a "UNION ALL" Neo4j adds a suffix
+        // "@<number>" to distinguish them,
+        //  so to check the correctness of the output parameters we must first remove this suffix from the column names
+        return keys.stream().map(i -> i.replaceFirst("@[0-9]+", "").trim()).collect(Collectors.toSet());
     }
 
     /*
@@ -392,9 +429,7 @@ public class Neo4jEmbeddingStore implements EmbeddingStore<TextSegment> {
     }
 
     private void createFullTextIndex() {
-        if (!fullTextAutocreate) {
-            System.out.println("Neo4jEmbeddingStore.createFullTextIndex");
-            
+        if (!autoCreateFullText) {
             return;
         }
